modulo-05-compose/app/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
gunicorn==21.2.0
